update sdk 060 版本

+++ include_lib/media/media_develop/media/audio_base.h
@@ -56,7 +56,7 @@ enum audio_channel {
 struct audio_fmt {
     u8  channel;
     u8  frame_len;
-    u16 sample_rate;
+    u32 sample_rate;
     u32 coding_type;
     u32 bit_rate;
     u32 total_time;
